Aulas.

+++ Curso_py/Desafio09.py
@@ -1,4 +1,4 @@
-valor = int(input('Digite o metro:'))
+valor = float(input('Digite o metro:'))
 centrimetro = valor * 100
 milimetro = valor * 1000
 print(f'O centrimetro é {centrimetro} e o milimetro é {milimetro}. ')
